requirements.txt
ccxt==4.3.75
pandas==2.2.2
python-dotenv==1.0.1
requests==2.32.3
